Add files via upload
The PAMELA (Papua Reports) application is a very valuable and important initiative to improve infrastructure problems in the Papua region. This application provides a platform for the Papuan people to report problematic infrastructure issues, such as damaged roads, collapsed bridges, or public facilities that are not functioning properly. With PAMELA, the Papuan people can actively participate in monitoring and reporting infrastructure problems in their area.